+++ scripts/extensions-emulator-tests/run.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 source scripts/set-default-credentials.sh
-./scripts/npm-link.sh
+./scripts/clean-install.sh
 
 (
   cd scripts/extensions-emulator-tests/functions